+++ docs/changelog.rst
@@ -11,6 +11,8 @@ Changelog
 news
 ++++
 
+- Adjust, for 3 and 7 categories, the method to calculate the contingency table in
+  percentage base on threshold size
 - New option in argument: -o --output dir, where 'dir' is the absolute and particular
   directory to save all results that makes Jaziku (this option is optional, default path
   for save result is the same where is the runfile)